move QbftConsensusConfig into core module
