+++ config/src/main/kotlin/maru/config/consensus/JsonFriendlyForksSchedule.kt
@@ -24,12 +24,12 @@ import kotlin.collections.component2
 import kotlin.collections.map
 import kotlin.collections.toSet
 import kotlin.reflect.KType
-import maru.config.consensus.qbft.DifficultyAwareQbftConfig
-import maru.config.consensus.qbft.QbftConsensusConfig
 import maru.consensus.ConsensusConfig
+import maru.consensus.DifficultyAwareQbftConfig
 import maru.consensus.ElFork
 import maru.consensus.ForkSpec
 import maru.consensus.ForksSchedule
+import maru.consensus.QbftConsensusConfig
 import maru.core.Validator
 import maru.extensions.fromHexToByteArray
