+++ config/src/test/kotlin/maru/config/consensus/JsonFriendlyForksScheduleTest.kt
@@ -9,11 +9,11 @@
 package maru.config.consensus
 
 import maru.config.MaruConfigLoader.parseBeaconChainConfig
-import maru.config.consensus.qbft.DifficultyAwareQbftConfig
-import maru.config.consensus.qbft.QbftConsensusConfig
+import maru.consensus.DifficultyAwareQbftConfig
 import maru.consensus.ElFork
 import maru.consensus.ForkSpec
 import maru.consensus.ForksSchedule
+import maru.consensus.QbftConsensusConfig
 import maru.core.Validator
 import maru.extensions.fromHexToByteArray
 import org.assertj.core.api.Assertions.assertThat
