+++ consensus/src/test/kotlin/maru/consensus/NextBlockTimestampProviderTest.kt
@@ -12,8 +12,6 @@ import java.time.Clock
 import java.time.Instant
 import java.time.ZoneId
 import kotlin.test.Test
-import maru.config.consensus.qbft.QbftConsensusConfig
-import maru.consensus.ElFork
 import org.assertj.core.api.Assertions.assertThat
 
 class NextBlockTimestampProviderTest {
