+++ serialization/src/test/kotlin/maru/serialization/ForkSpecSerializerTest.kt
@@ -9,10 +9,10 @@
 package maru.serialization
 
 import java.nio.ByteBuffer
-import maru.config.consensus.qbft.DifficultyAwareQbftConfig
-import maru.config.consensus.qbft.QbftConsensusConfig
+import maru.consensus.DifficultyAwareQbftConfig
 import maru.consensus.ElFork
 import maru.consensus.ForkSpec
+import maru.consensus.QbftConsensusConfig
 import maru.core.Validator
 import maru.core.ext.DataGenerators
 import org.assertj.core.api.Assertions.assertThat
